Update stats test for 2.15, fix SC-24594
